requirements.txt
requests==2.26.0
bs4==0.0.1
openai==1.51.0
praw==7.7.1
python-dotenv==1.0.1